+++ lib/dicey/distribution_calculators/empirical.rb
@@ -3,13 +3,14 @@
 require_relative "base_calculator"
 
 require_relative "../mixins/vectorize_dice"
+require_relative "../mixins/warn_about_vector_number"
 
 module Dicey
   module DistributionCalculators
     # "Calculator" for a collection of {AbstractDie} using empirically-obtained statistics.
     #
     # @note This calculator is mostly a joke. It can be useful for educational purposes,
-    #   or to verify results of {BruteForce} when in doubt. It is not used by default.
+    #   or to verify results of {BruteForce} when in doubt. It is not used automatically.
     #
     # Does a number of rolls and calculates approximate probabilities from that.
     # Even if weights are requested, results are non-integer.
@@ -20,6 +21,7 @@ module Dicey
     # - *rolls* (Integer) (_defaults_ _to:_ _N_) — number of rolls to perform
     class Empirical < BaseCalculator
       include Mixins::VectorizeDice
+      include Mixins::WarnAboutVectorNumber
 
       # Default number of rolls to perform.
       N = 10_000
@@ -30,10 +32,7 @@ module Dicey
         if defined?(VectorNumber) || dice.all?(NumericDie)
           true
         else
-          warn <<~TEXT
-            Dice with non-numeric sides need gem "vector_number" to be present and available.
-            If this is intended, please install the gem.
-          TEXT
+          warn_about_vector_number
           false
         end
       end