+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Next]
 
 **Added**
-- CLI can now be enabled with `require "dicey/cli"` instead of `require "dicey/cli/blender"`. CLI calls can be made with `Dicey::CLI.call([argv])`.
+- Key classes and modules now have documentation with important points and links to further reading to reduce friction for API users.
+- `Dicey::DistributionCalculators::AutoSelector.call` which does not require an instance to be created.
+- CLI is now enabled with `require "dicey/cli"` instead of `require "dicey/cli/blender"`. CLI is called with `Dicey::CLI.call([argv])`. CLI classes are considered private, however.
 
 **Changed**
 - Rename "--mode frequencies" option to "--mode distribution" and "--result frequencies" to "--result weights".